requirements.txt
streamlit>=1.32.0
pymupdf>=1.23.0
python-docx>=1.1.0
groq>=0.9.0
pillow>=10.0.0
